requirements.txt
langchain~=0.0.340
openai~=1.3.5
pinecone-client
tiktoken
promptwatch~=0.4.0
pinecone-text
langserve[all]~=0.0.30